2.8 - sistema de notificações completo
 - Adicionado suporte a NotificacaoProposta, NotificacaoCorretor e NotificacaoImobiliaria
- Criados DTOs específicos para cada tipo de notificação
- Implementadas respostas para solicitações com aprovar/recusar
- Integrado envio automático de notificações no fluxo de criação de proposta
- Refatorado PropostaService para delegar notificação ao NotificacaoService

+++ src/main/java/com/imveis/visita/Imoveis/repositories/NotificacaoPropostaRepository.java
@@ -5,10 +5,9 @@ import com.imveis.visita.Imoveis.entities.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface NotificacaoPropostaRepository extends JpaRepository<NotificacaoProposta, BigInteger> {
+public interface NotificacaoPropostaRepository extends JpaRepository<NotificacaoProposta, Long> {
     List<NotificacaoProposta> findByDestinatario(Usuario usuario);
 }
